refactor: move user picture methods to separate controller
